fix: consolidate auth middleware to _shared/middlewares.ts mixedAuth
